+++ src/algorithms/garp/rules_range.hh
@@ -54,7 +54,7 @@ public:
   
   virtual void initialize(GarpCustomSampler * sampler);
   virtual bool applies(Scalar * values);
-  virtual double getStrength(Scalar * values);
+  virtual int getStrength(Scalar * values);
 
   void log();
 };
